fixed auth issues
